Remove duplicate examples
There is a lot of code duplication between the examples, which makes it
hard to see the different approaches. The file based example is also
missing instructions how to obtain "sockex1".

Keep the example loading from a byte array, since it covers more parts
of the API.
